+++ saleor/account/emails.py
@@ -1,3 +1,5 @@
+from urllib.parse import urlencode
+
 from django.conf import settings
 from django.urls import reverse
 from templated_email import send_templated_mail
@@ -8,6 +10,18 @@ from ..core.emails import get_email_base_context
 from ..core.utils import build_absolute_uri
 
 
+def _send_password_reset_email(reset_url, recipient_email, user_id):
+    context = get_email_base_context()
+    context["reset_url"] = reset_url
+    send_templated_mail(
+        template_name="account/password_reset",
+        from_email=settings.DEFAULT_FROM_EMAIL,
+        recipient_list=[recipient_email],
+        context=context,
+    )
+    account_events.customer_password_reset_link_sent_event(user_id=user_id)
+
+
 @app.task
 def send_password_reset_email(context, recipient, user_id):
     reset_url = build_absolute_uri(
@@ -16,15 +30,17 @@ def send_password_reset_email(context, recipient, user_id):
             kwargs={"uidb64": context["uid"], "token": context["token"]},
         )
     )
-    context = get_email_base_context()
-    context["reset_url"] = reset_url
-    send_templated_mail(
-        template_name="account/password_reset",
-        from_email=settings.DEFAULT_FROM_EMAIL,
-        recipient_list=[recipient],
-        context=context,
-    )
-    account_events.customer_password_reset_link_sent_event(user_id=user_id)
+    _send_password_reset_email(reset_url, recipient, user_id)
+
+
+@app.task
+def send_password_reset_email_with_url(redirect_url, recipient_email, token, user_id):
+    params = urlencode({"email": recipient_email, "token": token})
+    reset_url = "%(redirect_url)s?%(params)s" % {
+        "redirect_url": redirect_url,
+        "params": params,
+    }
+    _send_password_reset_email(reset_url, recipient_email, user_id)
 
 
 @app.task
